+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### Adds
+
+* Adds pinia and a store to handle modals logic. Methods from the store are registered on `apos.modal` instead of component methods.
+
 ### Changes
 
 * Improves widget tabs for the hidden entries, improves UX when validation errors are present in non-focused tabs.
@@ -14,6 +18,10 @@ Now we want page B to be the child of page A.
 We will now end up with page B slug as `/peer/page` and not `/peer/peer/page` as before.
 * Set `G,Shift+I` shortcut to open the Image Tags manager modal.
 * Set `G,Shift+F` shortcut to open the File Tags manager modal.
+
+### Adds
+
+* Adds field components access to the reactive document value.
 
 ### Fixes